ml/requirements.txt
# L1 ML pipeline dependencies
# Install: pip install -r ml/requirements.txt
# Requires Python 3.11.9, numpy==1.26.4 (install first per CLAUDE.md)

# Data processing
polars>=1.0.0
pyarrow>=14.0.0

# Physiological signal processing
neurokit2>=0.2.9
cvxopt>=1.3.0          # required by NeuroKit2 cvxEDA decomposition

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.12.0

# ONNX
onnx>=1.16.0
onnxruntime>=1.18.0
skl2onnx>=0.5.0

# Experiment tracking
mlflow>=2.13.0

# Config + Excel
pyyaml>=6.0.1
openpyxl>=3.1.0
